requirements.txt
torch
transformers
datasets
Pillow
accelerate
torchvision